refactor: migrate diplomacy-editor to bundled TypeScript controller
Port public/modules/ui/diplomacy-editor.js to src/controllers/diplomacy-editor.ts
following the migration guide (strict d3 v7: select/pointer/interpolateString/
color named imports, event-arg map-click handler). The static #diplomacyEditor
chrome stays in index.html (rows generated at runtime); listeners wire once
behind an initialized flag. State 0's diplomacy field (the chronicle) is
cast to string[][] at the boundary. Register as Controllers.DiplomacyEditor
and replace editDiplomacy() at the tools.js and hotkeys.js call sites.
